refactor: separate lazy handlers of composer module
