+++ packages/composer/src/index.ts
@@ -1,4 +1,5 @@
-import { lazyComposer, lazyComposerSingle } from './composer/lazy'
+import { create, createSigle } from './composer'
+import { endHandler, errorHandler, executeHandler } from 'wezi-lazy-handlers'
 
-export const composer = lazyComposer
-export const composerSingle = lazyComposerSingle
+export const composer = create(endHandler, errorHandler, executeHandler)
+export const composerSingle = createSigle(errorHandler, executeHandler)
