+++ packages/lazy-handlers/.eslintrc
@@ -1,0 +1,3 @@
+{
+    "extends": "../../.eslintrc"
+}
