+++ packages/lazy-handlers/src/utils.ts
@@ -1,0 +1,3 @@
+export const isPromise = (obj: any) => typeof obj.then === 'function'
+export const isProduction = () => process.env.NODE_ENV === 'production'
+
