+++ tsconfig.json
@@ -17,6 +17,9 @@
       "wezi-error": [
         "packages/error"
       ],
+      "wezi-lazy-handlers": [
+        "packages/lazy-handlers"
+      ],
       "wezi-query": [
         "packages/querystring"
       ],
